Reorganize events livewire component
