+++ resources/views/livewire/events/events-list.blade.php
